requirements.txt
# Core libraries for Hoops Hustler
nba_api==1.8.0              # NBA stats pulling
ollama==0.4.7               # Local LLM integration
langchain==0.3.20           # Prompt and context chaining
langchain-community==0.3.19 # Community extensions for langchain
streamlit==1.43.2           # Web UI

# Key dependencies (pulled from your UV install)
numpy==1.26.4               # Data handling for nba_api
pandas==2.2.3               # Data manipulation
requests==2.32.3            # API calls in nba_api
pydantic==2.10.6            # Data validation in langchain
pyyaml==6.0.2               # Config handling in langchain
sqlalchemy==2.0.39          # Optional DB support (if we add SQLite)
tenacity==9.0.0             # Retry logic in langchain
httpx==0.28.1               # HTTP requests for ollama
jinja2==3.1.6               # Templating in streamlit
pydeck==0.9.1               # Visualization in streamlit
altair==5.5.0               # Charts in streamlit

# Additional dependencies
openai==1.66.3              # OpenAI API integration
tiktoken==0.9.0             # Tokenization for OpenAI
python-dotenv==1.0.1        # Load environment variables from .env

# LangChain Community dependencies (from your UV install)
aiohttp==3.11.14            # Async HTTP requests (langchain-community)
dataclasses-json==0.6.7     # JSON serialization (langchain-community)
marshmallow==3.26.1         # Data marshaling (langchain-community)
pydantic-settings==2.8.1    # Settings management (langchain-community)

# Testing dependencies
iniconfig==2.0.0            # Pytest configuration
pluggy==1.5.0               # Pytest plugin system
pytest==8.3.5               # Testing framework
tomli==2.2.1                # TOML parsing for pytest

# New dependencies
plotly==6.0.0
langchain-openai==0.3.8
newsapi==0.1.1